chore: Move code into crd module
