+++ crates/stackable-operator/src/commons/mod.rs
@@ -1,17 +1,14 @@
 //! This module provides common datastructures or CRDs shared between all the operators
 
 pub mod affinity;
-pub mod authentication;
 pub mod cache;
 pub mod cluster_operation;
-pub mod listener;
 pub mod networking;
 pub mod opa;
 pub mod pdb;
 pub mod product_image_selection;
 pub mod rbac;
 pub mod resources;
-pub mod s3;
 pub mod secret;
 pub mod secret_class;
 pub mod tls_verification;
