+++ crates/stackable-operator/src/constants/mod.rs
@@ -1,0 +1,1 @@
+pub mod secret;
