+++ crates/stackable-operator/src/constants/secret.rs
@@ -1,0 +1,1 @@
+pub(crate) const SECRET_BASE_PATH: &str = "/stackable/secrets";
