+++ crates/stackable-operator/src/crd/authentication/kerberos.rs
