+++ crates/stackable-operator/src/crd/authentication/static_.rs
