+++ crates/stackable-operator/src/crd/s3/mod.rs
@@ -1,0 +1,5 @@
+mod bucket;
+mod connection;
+
+pub use bucket::*;
+pub use connection::*;
