+++ crates/stackable-operator/src/lib.rs
@@ -4,6 +4,7 @@ pub mod client;
 pub mod cluster_resources;
 pub mod commons;
 pub mod config;
+pub mod constants;
 pub mod cpu;
 pub mod crd;
 pub mod helm;
